splitting psychopy functions into core and extra
